Abstract extractPropsFromComponentName
